Try renabling test on all platforms

+++ test/IRGen/Inputs/huge_c_type.h
@@ -5,7 +5,7 @@ typedef uint8_t bool;
 #define CREATE_ARRAY(T, N)                                                     \
   struct {                                                                     \
     T data[N];                                                                 \
-    size_t size;                                                               \
+    uint64_t size;                                                             \
   }
 
 typedef struct {
